+++ src/tree-sitter-grammar.js
@@ -1,6 +1,5 @@
-const path = require('path');
 const SyntaxScopeMap = require('./syntax-scope-map');
-const Module = require('module');
+const { loadLanguageModule } = require('./load-tree-sitter-language');
 
 module.exports = class TreeSitterGrammar {
   constructor(registry, filePath, params) {
@@ -50,15 +49,7 @@ module.exports = class TreeSitterGrammar {
       this.addInjectionPoint(injectionPoint);
     }
 
-    // TODO - When we upgrade to a new enough version of node, use `require.resolve`
-    // with the new `paths` option instead of this private API.
-    const languageModulePath = Module._resolveFilename(params.parser, {
-      id: filePath,
-      filename: filePath,
-      paths: Module._nodeModulePaths(path.dirname(filePath))
-    });
-
-    this.languageModule = require(languageModulePath);
+    this.languageModule = loadLanguageModule(params.parser, filePath);
     this.classNamesById = new Map();
     this.scopeNamesById = new Map();
     this.idsByScope = Object.create(null);